+++ native/packages/react-native-bpk-component-banner-alert/src/BpkBannerAlert.android.js
@@ -44,6 +44,7 @@ import {
 import BpkAnimateHeight from 'react-native-bpk-component-animate-height';
 import BpkIcon, { icons } from 'react-native-bpk-component-icon';
 import BpkText from 'react-native-bpk-component-text';
+import BpkTouchableNativeFeedback from 'react-native-bpk-component-touchable-native-feedback';
 
 import AnimateAndFade from './AnimateAndFade';
 import ALERT_TYPES from './AlertTypes';
@@ -121,7 +122,7 @@ const ALERT_TYPE_STYLES = {
 const DismissableButton = ({ dismissButtonLabel, onDismiss }) => (
   <View style={styles.dismissableContainer}>
     <View style={styles.dismissButtonContainerRadius}>
-      <TouchableNativeFeedback
+      <BpkTouchableNativeFeedback
         accessibilityComponentType="button"
         accessibilityLabel={dismissButtonLabel}
         background={TouchableNativeFeedback.SelectableBackgroundBorderless()}
@@ -130,7 +131,7 @@ const DismissableButton = ({ dismissButtonLabel, onDismiss }) => (
         <BpkText textStyle="sm" emphasize style={styles.dismissLabel}>
           {dismissButtonLabel.toUpperCase()}
         </BpkText>
-      </TouchableNativeFeedback>
+      </BpkTouchableNativeFeedback>
     </View>
   </View>
 );